+++ src/main/kotlin/instance/Instance.kt
@@ -25,10 +25,14 @@ data class Instance(
     val state: InstanceState,
     val uses4GBPatch: Boolean,
     val isBaseInstance: Boolean = false,
-    val icon: Path? = null
-) : PersistentValue<SavedInstance> {
+    val icon: Path? = null,
+    val baseInstance: SerializableUUID? = null
+) : PersistentValue<SavedInstance>, Translatable {
     val manifest: Path get() = directory / "manifest.acf"
     val manifestBackup: Path get() = directory / "manifest.backup.acf"
+
+    override val translation: Strings.() -> String
+        get() = { if (isBaseInstance) baseInstance else name }
 
     suspend fun start(editor: Boolean = false, awaitSteamDeath: suspend () -> Boolean = { true }) {
         val flags = buildList {
@@ -64,12 +68,49 @@ data class Instance(
     }
 
     override fun toSavedData(): SavedInstance {
-        return SavedInstance(id, name, directory, patchVersion, options, uses4GBPatch, isBaseInstance, icon = icon)
+        return SavedInstance(
+            id,
+            name,
+            directory,
+            patchVersion,
+            options,
+            uses4GBPatch,
+            isBaseInstance,
+            icon = icon,
+            baseInstance = baseInstance
+        )
     }
 }
 
 val List<Instance>.baseInstance: Instance
     get() = firstOrNull { it.isBaseInstance } ?: error("Base instance not found")
+
+fun List<Instance>.calculateReLinkRoundsFor(instance: Instance): Map<Instance, List<Instance>> {
+    val directDependencies =
+        filter { it.baseInstance == instance.id || (instance.isBaseInstance && it.baseInstance == null && !it.isBaseInstance) }
+    val childDependencies = directDependencies.map {
+        it to calculateReLinkRoundsFor(it).values.flatten()
+    }.filterNot { (_, children) -> children.isEmpty() }
+
+    return buildMap {
+        this[instance] = directDependencies
+        putAll(childDependencies.toMap())
+    }
+}
+
+fun List<Instance>.findDependenciesFor(instance: Instance) = filter {
+    it.hasDependencyTo(instance, this)
+}
+
+fun Instance.hasDependencyTo(instance: Instance, instances: List<Instance>): Boolean {
+    if (baseInstance == null) return instance.isBaseInstance && !isBaseInstance
+    val parent = instances.first { it.id == baseInstance }
+    return if (parent.id == instance.id) {
+        true
+    } else {
+        parent.hasDependencyTo(instance, instances)
+    }
+}
 
 @Serializable
 enum class LaunchFlag(val cliName: String) {