+++ src/main/kotlin/localization/Strings.kt
@@ -1,6 +1,7 @@
 package dev.nycode.omsilauncher.localization
 
 import androidx.compose.ui.text.AnnotatedString
+import dev.nycode.omsilauncher.instance.Instance
 
 data class Strings(
     val ok: String,
@@ -20,6 +21,7 @@ data class Strings(
     val yes: String,
     val no: String,
     val confirmDeletion: (instance: String) -> String,
+    val confirmDeletionWithDependencies: (instance: String, dependencies: List<Instance>) -> String,
     val instanceName: String,
     val activateInstance: String,
     val activate: String,
@@ -82,7 +84,7 @@ data class Strings(
     val waitingForChanges: String,
     val iAmDone: String,
     val mergingSteamManifest: (name: String) -> String,
-    val reLinkingInstances: String,
+    val reLinkingInstances: (current: Int, total: Int) -> String,
     val save: String,
     val editInstanceTitle: (name: String) -> String,
     val editBaseInstance: String,