+++ src/main/kotlin/localization/StringsEn.kt
@@ -27,6 +27,14 @@ val StringsEn = Strings(
     confirmDeletion = { instance ->
         "Are you sure you want to delete $instance?"
     },
+    confirmDeletionWithDependencies = { instance, dependencies ->
+        "Are you sure you want to delete $instance and the following instances: ${
+            dependencies.joinToString(
+                "\n",
+                prefix = "\n",
+            ) { " - ${it.name}" }
+        }"
+    },
     instanceName = "Instance name",
     activateInstance = "Activate instance",
     activate = "Activate",
@@ -98,7 +106,7 @@ val StringsEn = Strings(
     iAmDone = "I am done",
     mergingSteamManifest = { name -> "Merging Steam manifest of $name" },
     preparingReSynchronisation = "Preparing re-synchronisation",
-    reLinkingInstances = "Relinking instances",
+    reLinkingInstances = { current: Int, total: Int -> "Relinking instanced, round $current/$total" },
     waitingForChanges = "Waiting for changes",
     save = "Save",
     editInstanceTitle = { name -> "Edit instance $name" },