+++ src/main/kotlin/ui/instance/context/modification/InstanceModificationState.kt
@@ -6,12 +6,13 @@ import androidx.compose.runtime.setValue
 import dev.nycode.omsilauncher.instance.Instance
 import dev.nycode.omsilauncher.instance.InstanceOptions
 import java.nio.file.Path
+import java.util.UUID
 import kotlin.io.path.exists
 
-class InstanceModificationState(parentInstance: Instance? = null) {
+class InstanceModificationState(parentInstance: Instance? = null, isEdit: Boolean = false) {
     var icon: Path? by mutableStateOf(if (parentInstance?.icon?.exists() == true) parentInstance.icon else null)
-    var name: String by mutableStateOf(parentInstance?.name.takeIf { parentInstance?.isBaseInstance == false } ?: "")
-    var path: Path? by mutableStateOf(parentInstance?.directory.takeIf { parentInstance?.isBaseInstance == false })
+    var name: String by mutableStateOf(parentInstance?.name.takeIf { parentInstance?.isBaseInstance == false || isEdit } ?: "")
+    var path: Path? by mutableStateOf(parentInstance?.directory.takeIf { parentInstance?.isBaseInstance == false || isEdit })
     var customPath: Path? by mutableStateOf(null)
     var patchVersion: Instance.PatchVersion by mutableStateOf(
         parentInstance?.patchVersion ?: Instance.PatchVersion.BI_ARTICULATED_BUS_VERSION
@@ -25,4 +26,5 @@ class InstanceModificationState(parentInstance: Instance? = null) {
     var screenMode: InstanceOptions.ScreenMode by mutableStateOf(
         parentInstance?.options?.screenMode ?: InstanceOptions.ScreenMode.DEFAULT
     )
+    var parentInstance: UUID? by mutableStateOf(parentInstance?.id)
 }